E(distace) for Mo_S2_V2 with default and optimzed parameters.

+++ scripts/ASPIRE-1/run_stability_trials.sh
@@ -3,7 +3,7 @@ TRIAL_NAME=$1
 TARGET=$2
 REPEATS=$3
 BATCH_SIZE=$4
-EXPERIMENT_NAME="combined_mixed_weighted_test"
+EXPERIMENT_NAME=$5
 
 if [ $((REPEATS % BATCH_SIZE)) -ne 0 ]; then
     echo "REPEATS must be a multiple of BATCH_SIZE"
@@ -17,7 +17,10 @@ PROJECT_ROOT=$(realpath "$SCRIPTPATH/../../")
 TRIALS_PATH=$PROJECT_ROOT/"trials"
 mkdir -p $TRIALS_PATH/stability/$TRIAL_NAME
 for i in $(seq 1 $REPEATS); do
-    cp $TRIALS_PATH/$TRIAL_NAME.yaml $TRIALS_PATH/stability/$TRIAL_NAME/$i.yaml
+    # Since for CatBoost we manually change the random seed, don't overwrite the present files
+    if [ ! -f $TRIALS_PATH/stability/$TRIAL_NAME/$i.yaml ]; then
+        cp $TRIALS_PATH/$TRIAL_NAME.yaml $TRIALS_PATH/stability/$TRIAL_NAME/$i.yaml
+    fi
     # Check if we have a batch of trials to process
     if [ $((i % BATCH_SIZE)) -eq 0 ]; then
         TRIALS_BATCH=$(seq -f stability/$TRIAL_NAME/%g $((i - BATCH_SIZE + 1)) $i)
